fully switch to GitHub Actions for CI
We were only using it now for fuzzit, but we've moved that to GitHub
Actions as well, setting up the required secret.

Also enable Windows on Actions, with a TODO to reenable the troublesome
test once they fix their %TEMP% bug.

Finally, update the fuzzit dependency, since it now knows to use Actions
env vars to fetch the git commit sha and branch.