cmd/shfmt: stop relying on the system's diff tool
Thanks to Josh Bleecher Snyder's excellent pkg/diff module, which is
able to produce a unified diff in pure Go. Since the API is also quite
simple, we end up with considerably less lines, and we are able to have
the feature working independently of the system.

This slightly changes some of the diffs that we produce, but that's
fine, as they are still valid. More importantly, the output is now
predictable and under our control.

This change does add an external dependency, but that's fine; it's a
small one, and we already relied on x/crypto anyway.