interp: refactor ModuleExec and separate LookPath
As pointed out by Konstantin Kulikov, performing the LookPath before
running an ExecModule is wasteful, because perhaps it's not needed.

Remove the argument from the func type, which simplifies the vast
majority of the tests and examples. For those that do need to find an
executable, export LookPath.

To not add a dependency on Runner, completely separate LookPath from it.
All it really needs is Env; it can fetch the current directory via PWD
too.

Fixes #413.