interp: improve the ModuleOpen API
ModuleOpen received an absolute path, which is unnecessary, since the
context already contains the current directory. Moreover, this meant
that we needed UnixPath to then implement certain modules, such as the
one to implement a portable /dev/null.

Fix this, and remove UnixPath. Thanks to Konstantin Kulikov for the
suggestion.

Updates #414.