interp: fix the tests on Alpine, and as root
The KillTimeout tests can be written in POSIX Shell, to not require
Bash.

A few tests assumed a non-root user, so make them pass when ran as root
as well.

To ensure that we support Alpine on Docker, add a CI job for it. This
also tests the lack of cgo, running as root, and the lack of bash, gnu
coreutils, glibc, and so on.

Updates #416.