interp: replace ShellExitStatus with a method
Needing to know whether the interpreter just exited the entire shell
session or program is only necessary in some use cases. In many others,
such as when calling Run exactly once, it's not relevant at all. The
non-nil error ShellExitStatus(0) could be confusing to users.

Instead, expose this information via an opt-in Exited boolean method. As
showin in cmd/gosh, it's even simpler to use than the old error type
switch mechanism.

Fixes #417.