pattern: split API from the syntax package
The syntax package already has a lot of code and exposed API. Moreover,
pattern matching syntax isn't directly part of the shell syntax, so it
doesn't need to be in the same place.

A separate package is less confusing for the users, allows for a much
nicer and small godoc page for the new package, and will enable us to
improve its API without worrying about syntax.

Take the opportunity to name the exported APIs better, too. Thanks to
Roger Peppe for his input.