interp: add a simple Interactive method
Mimicking Parser.Interactive, this method implements the basics of a
full interactive shell. That's useful for many user-facing use cases,
and also as a debugging tool.

The function has zero options because the list of options it could take
is never-ending. We could start with PS1 and PS2, but we could also add
PS3, PS4, a context.Context, Parser options, et cetera.

Instead of tripling the amount of code and adding lots to the API,
suggest that advanced users simply copy the twenty lines to add their
own options and logic.