cmd/shfmt: add initial EditorConfig support
Using the editorconfig module we've written over the past weeks.

See the README for a bit of docs on how it works. In particular, one can
entirely disable the feature by formatting stdin or using any formatting
flag.

We don't do any merging of flags from EditorConfig and the command line
right now; this keeps the implementation simple, and allows formatting
entire directories while skipping EditorConfig files with no-op flags
like "shfmt -ln=bash src/".

It's not clear that the ability to merge options from both sides would
be desirable; we might reconsider in the future.

We've also added an undocumented SHFMT_NO_EDITORCONFIG environment
variable, which can be set to "true" to completely disable the feature
in case it breaks badly when first released. This flag will be removed
once the feature has been released and stable for a few months.

Fixes #393.