cmd/shfmt: use EditorConfig if no args are given too
As pointed out by a user, using standard input is common within IDEs and
editors, which should obey EditorConfig too.

It's always possible to skip the feature via a flag. Add an example to
make that clearer.

The change also makes the feature simpler, as now there's just one way
to skip it entirely, instead of two.

Fixes #450.