shell: remove Source* funcs
They were a bit basic and not very configurable. For example, they
didn't allow for any parser or interpreter options, and stdout/stderr
were entirely ignored.

We could redo the APIs to make them more powerful, but there's probably
little point. If someone wants that much customization, they could write
the dozen lines of code themselves and have all the power.

The only internal detail from the removed code are the predefined
variables like $HOME and $PWD. We might re-expose this somehow in the
interp package, as it could be useful in general.

Fixes #427.