syntax: stop parsing comments at escaped newlines
We correctly kept the backslash and newline as part of the comment
value, but we didn't stop the lexer. This meant that, if a comment ended
in a backslash, we'd include the next line as part of the same comment.

Unfortunately, there was no test case for this edge case. Add one.

Fixes #457.