interp: add support for <( process substitutions
The >( form doesn't work yet. It requires careful coordination between
the main runner and the nested concurrent runner, as it's the parent
that must close the pipe and wait for the child to finish.

I haven't quite figured out a clean way to write that. Leaving it as a
TODO for now.

Updates #451.