syntax: support for loops with braces on bash/mksh
We support these like we support command substitutions with backquotes;
as a deprecated syntax. The parser accepts these, but the printer never
outputs the deprecated form.

The feature is documented as a historical feature in "man mksh", and it
has worked on Bash forever even though it's undocumented. It has never
been documented in the POSIX Shell spec, nor does it work on popular
shells like dash and ash. Error if the syntax is used in POSIX mode.

Fixes #444.