interp: make two tests work on Bash 5.0
4393a3a0a4 was very helpful in that it added tests. However, these two
don't work on vanilla Bash.

The first tries to use associative arrays. That works with our package,
because we assume that quoted indexes mean that. However, bash still
interprets that as an indexed array. Thus, x[x]=x is recursive forever,
and makes bash fail.

The second tries to test that changing an alias in a subshell doesn't
affect the parent shell. However, because Bash parses the entire
subshell source before executing it, the parent's alias applies before
we modify the value of the alias when executing. This is not happening
with our shell, because alias expansion happens after parsing.

To fix the second problem, add an eval, so that we can force Bash to
expand the alias at execution time.

This was unfortunately not caught by CI, as it doesn't have a new enough
Bash. But I do.