interp: handle the last exit status code better
In some edge cases, such as commands with no arguments or background
commands, we didn't properly reset it to 0.

Doing that breaks expansions like "$?"; to properly handle all cases, we
need to keep track of the current and last status codes separately, at
least while we execute a statement.

As a bonus, we can deduplicate a few places where we set r.exit = 0.

Fixes #506.