interp: don't use global file descriptors in test -t
Instead, support IsTerm and Fd methods on the interface values. The
latter is supported by *os.File, and the former can be implemented like
we do in our test.

In the future, access to the global file descriptors will be possible
when the "single process mode" is used, as described in #171. For now,
we never go that route.

Fixes #508.