interp: test tty detection properly
This includes two table-driven tests to detect whether FDs 0-3 are
terminals. One uses the Go API with StdIO, without spawning a separate
process, and another uses os/exec to run the interpreter in a separate
process via TestMain.

The cases are very similar, except that with pure Go we can fake a
buffer to be a terminal via the IsTerm method. The exec method also has
real FDs 0/1/2, versus the native version's, whose global FDs 0/1/2
belong go 'go test'.

We remove the IsTerm interface too, as it was initially added mainly for
testing, and it's no longer strictly needed. We can re-add it in the
future as part of the API, if we think it's useful.

We also add github.com/creack/pty, so that we can test a pseudoterminal
in both cases. This isn't supported on Windows, so don't run the tests
there.

While at it, switch from x/crypto to x/term for IsTerminal. It's the new
home for this API, and the module is far smaller.

Finally, a few changes to the 'test -t' logic needed to be made to
properly support all of the combinations above. The mistakes in the old
code were correctly spotted by Larry Clapp.

Thanks to Larry once again for helping get the pty tests working on Mac.

Fixes #508, again.