cmd/shfmt: atomically replace files with -w
First, add a test case where a script uses shfmt to format itself.
Before this change, we'd get a failure, since the file that the shell is
executing gets modified in-place:

	> exec sh input.sh
	[stdout]
	foo
	[stderr]
	input.sh: line 3: ho: command not found

The test requires a bit of magic to have the shfmt tool in $PATH without
installing it globally, nor building it every time. We reuse a method
form another tool which simply reuses the test binary.

Use renameio to instead replace the input file atomically. One open
question is file metadata; because the old file is thrown away, we would
normally lose its permissions and owner/group. Do our best to keep the
permission bits in place via a separate lstat call.

The separate lstat call isn't atomic, but that's okay. If something is
deleting or messing with the file while shfmt is running, other parts of
the tool could fail already, such as parsing it. Keeping the permissions
in-place is important, for example to preserve the original script's
executable bit, independently of whether it was set or unset.

For now, this feature isn't enabled on Windows, since renameio doesn't
support the operating system. On it, we fall back to the previous
ioutil.WriteFile, which isn't an atomic write, but still works.

Fixes #401.