syntax: support disabling KeepPadding
This was a big deal for shfmt's support for editorconfig files, since it
relies on the ability to enable or disable options as per the currently
applied editorconfig properties.

There were two bugs here; the first, found by a user, was that enabling
the option twice would crash. The second, which was silent and more
dangerous, was that disabling the option did nothing. The latter could
result in incorrect formatting options being used for some files.

Fixes #530.