CI: stop building linux/arm/v6 docker images
Go's debian-based docker images don't support that platform.

I doubt there are many docker users on armv6. The few who might be
affected can probably download a binary, or spend ten seconds to do a
build.

We used to use the alpine-based image, but that lacks git, so the builds
required extra work. That seems like a good change we don't want to
lose.