syntax: keep trailing escaped newlines in heredocs
v3 started representing escaped newlines via position information.

Since the stop word of a heredoc doesn't have position information, we
must record the proper end position of the heredoc via the last part of
its word. So, if the body ends with escaped newlines, add an empty
literal with the right position.

Fixes #557.