interp: use os.Pipe instead of io.Pipe for shell pipes
Initially, there wasn't a reason we chose io.Pipe instead of os.Pipe.
Both seemed to work equally fine for our needs.

However, io.Pipe had a fatal flaw; since the ends are just an io.Reader
and io.Writer in Go, and they are not OS files, they cannot be used
directly as stdin/stdout/stderr when executing programs in forked
processes.

This was fine for stdout/stderr; they still behaved fine, even though
os/exec had to do an extra os.Pipe call for each, to properly make the
writer accessible to the child process. An extra goroutine was also
needed to do the io.Copy on the Go side.

However, the story was worse for stdin, which used the same os.Pipe plus
io.Copy mechanism. io.Copy starts reading from the source even if the
destination isn't ready to have bytes written to. As a result, an
executed program would start consuming stdin even if the program in
question didn't issue a single stdin read.

For instance, the following program would print nothing, as the first
cat would consume the four "foo\n" bytes, even though it did nothing at
all with them.

	echo foo | { cat /dev/null; cat; }

The extra interaction with the OS adds a small overhead, but we make a
big correctness gain, and we also allocate much less garbage. In
particular, we can throw away bufCopier entirely, as we can simply rely
on the kernel's buffering.

	name   old time/op    new time/op    delta
	Run-8    1.71ms ± 1%    1.76ms ± 0%   +2.69%  (p=0.000 n=8+8)

	name   old alloc/op   new alloc/op   delta
	Run-8     186kB ± 0%      84kB ± 0%  -54.75%  (p=0.000 n=8+7)

	name   old allocs/op  new allocs/op  delta
	Run-8       784 ± 0%       748 ± 0%   -4.59%  (p=0.000 n=8+7)

Fixes #566.