expand: support globbing windows short path names
They are similar to symlinks; one path that, when opened, is actually
another path.

The difference is that, when listing the parent directory of a symlink,
we will find said symlink. This is not the case with windows short
paths; they are an abbreviation done by the system, they do not exist on
the filesystem.

This was breaking globbing, because we expected to find each path
element on the filesystem. Instead, if an element has no globbing
characters, we simply stat it instead of listing its parent directory to
find it.

While this increases the amount of code, it also means less work for the
general case where a path element contains no glob characters. We end up
doing slightly more work here, because this patch is mainly a bug fix.
However, we leave a TODO since there's plenty of room to optimize.

	name   old time/op    new time/op    delta
	Run-8    1.70ms ± 0%    1.71ms ± 1%  +0.54%  (p=0.041 n=6+6)

	name   old alloc/op   new alloc/op   delta
	Run-8    83.9kB ± 0%    84.7kB ± 1%  +1.03%  (p=0.002 n=6+6)

	name   old allocs/op  new allocs/op  delta
	Run-8       744 ± 0%       752 ± 0%  +1.08%  (p=0.008 n=5+5)

Finally, add a test under interp which properly tests against a real
symlink on non-Windows, and against a real short path on Windows.