interp: make pid_and_hang use a sleep
Because we have a single goroutine, using an empty select can actually
trigger an "all goroutines are asleep" deadlock panic.

Use a very long sleep, which works just as fine for our needs, and won't
cause a deadlock.

We spotted these crashes by grabbing stdout from the call, so leave that
code in to help future debugging. It's not clear why this panic happened
more often with GOARCH=386.

Fixes #573.