syntax: stop heredocs when the entire line matches
Otherwise, with a heredoc starting like <<EOF, a line like "EOF_foo"
would be mistaken for the end of the entire heredoc. This is because we
would only do a prefix match.

The reason for the prefix match was because the "current line" might or
might not contain a newline. Instead, remove any trailing newline, and
do a bytes.Equal comparison.

Add tests too, for both the regular and quoted heredoc cases.

Fixes #575,