cmd/shfmt: add -filename to name standard input
Otherwise, EditorConfig support is pretty useless with stdin, as we
would always use "<standard input>" as the filename.

This feature is particularly useful for editor integrations, which want
to format a real file on disk but with the in-memory contents that the
user is editing.

Start writing the changelog for 3.2.0 too.

Fixes #577.