syntax: improve parse errors for missing test expressions
Before, many of the added test cases would error with messages including
"must be followed by a word", even though && is a binary test operator
which can accept any test expression to its right, not just a word.

This was due to testExprBase emitting the "must be followed by" error
directly, but that was incorrect as it doesn't have that nuance. Some
operators like -e do require a word and not just any expression, but
that's just part of them.

To fix this issue, make testExpr return an untyped nil when we reach EOF
or a closing token, and then the caller will decide what "must be
followed by" error is appropriate.

For #590.