syntax: properly escape tabs in tab-indented heredocs
We use text/tabwriter to vertically align some nodes like comments.
Because of that, we need to escape all indentation tab characters with
tabwriter.Escape, since they don't concern those alignments.

We weren't doing this properly for the <<- heredocs. We were printing
the heredoc body itself with the escape characters, and we'd try (and
fail) to amend them later when we add the extra indentation.

Instead, don't add those if we're the nested printer being used for the
heredoc body. If both layers of printers add the escape bytes, they'll
just end up cancelling each other out on occasion.

Fixes #584.