interp: $@ and $* always exist
The interpreter had a bug where the empty expansion is not removed in a
call when a shift operation caused it to become empty. This is due to
shift's behavior of setting interp.Runner.Params to nil when no more
positional parameters are left.

Unfortunately, looking up $@ and $* returns a value with with List set
to interp.Runner.Params, meaning the List field can be nil when the
variable claims to be set. This breaks expand's expectation of List type
existence meaning non-zero values are available, as seen in
`quotedElemFields`.

This changes interp's environment lookup to always return a non-nil
[]string for $@ and $*, satisfying expand's expectations.

Fixes #606.