syntax: add support for Bats
Based in https://github.com/bats-core/bats-core, Bats is a testing
framework for Bash. It's really just Bash with one single syntax
addition to allow defining tests, such as:

	@test "addition using bc" {
		result="$(echo 2+2 | bc)"
		[ "$result" -eq 4 ]
	}

Add support for it, which is a thin layer on top of our Bash support.

No changes are made to the interpreter, because we don't intend to
actually implement running those tests for now. The point of supporting
the language in the syntax package is for formatting in shfmt.

Fixes #600.