cmd/shfmt: always format explicitly given files
We simplified the directory walking code in 9096347a87, with the intent
to deduplicate code. Deduplicating the code to handle shebangs was
indeed a good idea, but we introduced a regression that wasn't covered
in the tests.

If a path to a file is given explicitly, and the -f flag isn't being
used, we should format that file even if it has neither a shell filename
extension nor a shebang.

Fixes #600 again, since "bats" is not considered a shell extension just
yet.