cmd/shfmt: fix regression with the -ln default of "bash"
In dfe3c9b150 there was a regression where, if any other formatting
flags were given, we'd go into "no editorconfig" mode and apply an
invalid language variant to the parser. This is because we introduced a
regression where "bash" was no longer the default when the flag is
unset.

Fix that, and add a test. syntax.Variant now also rejects invalid
LangVariant values via a panic, similar to other options like StopAt.
This would have caught our mistake early on.

	panic: unknown shell language variant: -1

Fixes #617.