syntax: don't simplify ${name[i]} into name[i]
We simplify simple parameter expansions within arithmetic expressions
when the Simplify flag is used.

This included indexes, but that was wrong because many common parameter
expressions like ${name[@]} would break; name[@] is not valid in bash or
ksh. zsh does accept it, but that's far from portable.

Just stop simplifying parameter expansions which include indexes. We
lose some valid simplifications like ${name[0]}, but it's better to be
conservative and consistent.

Fixes #616.