syntax: make parser guard line/column overflows
After this change, reaching past the maximum line or column numbers
results in zero line and column numbers reported by positions, which is
clearly documented in the godoc.

This way, such numbers can clearly be treated or rendered as invalid or
unknown, as opposed to wrapping around to 0 and being confusing to the
end user.

In particular, Pos.String now prints such unknown numbers as "?".

We also amend Pos.IsValid to work properly with unknown line numbers, by
instead relying on the offset as well. This fixes a problem in the
parser where we'd rely on Stmt.Semicolon.IsValid to check if a statement
had a semicolon end token, which would incorrectly return false if the
line number wrapped around to 0.

This protection costs us about 2% of the overall parsing time, but I
reckon it's worth it for the added safety and user-friendliness.

	name     old time/op    new time/op    delta
	Parse-8    27.1µs ± 1%    27.6µs ± 0%  +1.83%  (p=0.004 n=6+5)

	name     old alloc/op   new alloc/op   delta
	Parse-8    14.7kB ± 0%    14.7kB ± 0%    ~     (all equal)

	name     old allocs/op  new allocs/op  delta
	Parse-8      99.0 ± 0%      99.0 ± 0%    ~     (all equal)

Updates #620.