cmd/shfmt: centralise detailed docs in a man page
Using https://sr.ht/~sircmpwn/scdoc, which is extremely similar to ronn
but does not add Ruby as a build dependency. The nice property is that
the manpage is basically Markdown, which allows us to easily read and
write it, but also allows decent direct rendering on GitHub.

Centralizing the detailed documentation for shfmt in the manpage is also
nice, because there was a bit of a tug of war between the entire
project's README.md and 'shfmt -h' for that role. Now, both contain very
basic introductions depending on the context, and then link to the
manpage.

Another advantage is that a lot of these extra details were only in the
README before, which is often not shipped in binary distributions. In
comparison, 'man shfmt' should soon work on most shfmt installations.

Fixes #574.