syntax: don't move inline heredoc comments to previous lines
The input:

	while foo; do
		bar
	done <<-EOF # comment
		baz
	EOF

Would get misprinted as:

	while
		# comment
		foo
	do
		bar
	done <<-EOF
		baz
	EOF

This is because we would consider "#comment" to be in the middle of the
statement, and print it too early. This is technically right, but
doesn't work well with heredocs.

Instead, change the logic to make this separation at the end of the
inner command, which excludes any redirections.

Fixes #623.