syntax: decrement indentation level when ending [[
In f9bde4fea9, we overhauled printing of test expressions to not require
escaping newlines. This was to fix #565.

However, we made a mistake; we forgot to decrement the indentation level
at the end of the test expression, to compensate for the inicrement at
the beginning.

This could result in leftover indentation after a test expression is
printed. For example, the input:

	{
		[[ foo ]]
	}

Would misprint as:

	{
		[[ foo ]]
		}

Fix that and add a regression test.

Fixes #628.