syntax: add the SingleLine option
This can be used to print simple bits of shell into a single line. The
first use-case is for printing shell command lines for interactive
guides. In that scenario, printing:

	$ (cd some-dir; foo bar baz)
	output of foo bar baz

is much easier than showing:

	$ (
	>	cd some-dir
	>	foo bar baz
	>)
	output of foo bar baz

The multi-line form is best for shell scripts and maintainable source,
but not for one-liners meant for copy-pasting or demonstrating single
operations.

A similar option is Minify, but it serves an entirely different purpose.
They both affect how newlines are printed, but Minify has no intent to
reduce the number of newlines to zero - it merely makes the output as
small as possible.

For now, Minify and SingleLine can't be used together, since they likely
don't make sense. If someone has a use case for this, they can file an
issue via the supplied error message.

shfmt does not currently expose this option, because it doesn't seem
terribly useful from the command line. Its initial use case comes
directly from using the syntax package as a Go library. We might
consider exposing the option to shfmt in the future, like Minify.