interp: support complex parameter expansions with nounset
When the "nounset" shell option is set, we were erroring out on unset
variables too early, at the level of the interpreter's os.Getenv
equivalent.

This does not work well, because in many cases, a parameter expansion
does not trigger an error even though it initially fetches an unset
variable:

	$ set -u; echo ${unset:-default}
	default

To fix this problem, we need to catch unset variables at a later stage,
while we're expanding parameter expansions. Since this happens in the
expand package, its Config needs to gain a NoUnset boolean parameter.

Add plenty of tests, too.

Thanks to Jonas Chevalier for raising this bug and providing the
TestRunnerOpts test case.

Fixes #627.