interp: unify code to parse builtin command flags
We had written the fancier flag parser to support "-abc" and not just
"-a -b -c" for the "set" builtin, but some other builtins were still
using the much simpler (and incorrect) inline parser.

Don't reuse the getopts parser, because its API is very awkward and it
doesn't support "+a" flags, needed by some builtins like "set".

Thanks to Jonas Chevalier for the suggestion.