interp: implement "type -t"
This flag is mostly used by scripts that do introspection.

Mark all the missing 'type' builtin options as NOT IMPLEMENTED. Just to
be explicit about what is missing.


Also add IsKeyword to the syntax package, to support keywords in the type
builtin. This is not entirely accurate as the answer depends on the parser
running mode. But it's a start.