re-implement and improve fuzz option encoding
We were using the first byte to encode 8 bits of options. The code was
hard to understand and maintain, and it omitted many new options like
SingleLine.

Use three bytes instead, which covers more ground. And use constants to
better express what each bit means.

This should mean that fuzzing is slower, but ultimately more powerful as
we cover more edge cases.