syntax: use more bits for position line numbers
The Pos type, used in nodes to report positions, used 32 bits for the
byte offset, 16 bits for the line number, and 16 bits for the column
number.

Unfortunately, this meant a limit of just 65k lines, after which
positions would render as "?" overflows. That is a low limit, compared
to the other limits in place; 4GiB offsets, and 65k columns.

Moreover, line numbers are generally more important in source code than
column numbers. Only showing a line number to the user is more useful
than only showing a column number. It is also fairly simple to figure out
the column number via the offset; simply count the number of bytes up to
the previous newline.

As such, shift two bits from the column numbers to the line numbers.
Now, the maximums supported before an overflow are 262k lines, and 16k
columns. We can call this a 18:14 split, where the original was a 16:16
split.

The machinery for arbitrary splitting of the 32 bits between line
numbers and column numbers is now in place, so it would be trivial to
tweak these numbers in the future. For example, we could also consider a
20:12 split, where we'd reach past one million lines, but limit column
numbers to just 4k.

The added bit shifting is cheap but it's done for every token, so it
does add a tiny amount of overhead. Better position information seems
worth the trouble.

name     old time/op    new time/op    delta
Parse-8    28.3µs ± 0%    28.7µs ± 0%  +1.22%  (p=0.001 n=6+7)

name     old alloc/op   new alloc/op   delta
Parse-8    14.7kB ± 0%    14.7kB ± 0%    ~     (all equal)

name     old allocs/op  new allocs/op  delta
Parse-8      99.0 ± 0%      99.0 ± 0%    ~     (all equal)

Fixes #620.