interp: implement trap of ERR and EXIT signals
Thanks to Jonas Chevalier for implementing the bulk of this on
https://github.com/mvdan/sh/pull/650.

His changes were mainly implementing the "trap" builtin itself,
recording the trap callbacks in new fields, and refactoring the exit
logic to allow calling those callbacks. And, of course, the tests.

My changes on top extend the tests a bit, make Runner.stmtSync call
the ERR trap callback when a simple command fails, and make Runner.Run
call the EXIT trap callback when done interpreting an entire *File.

That last change is backwards compatible, since the only noticeable
effect should be that Run on a *File can call EXIT trap callbacks. Those
were not supported before, so no existing program should break.

Finally, note that we add a test case that would make the Runner get
stuck in a callback loop without the handlingTrap boolean field.

Fixes #635.