cmd/shfmt: document ways to enable and disable color
The tests used FORCE_COLOR to explicitly enable colored diffs in the
tests, even though they don't have a terminal output. Some users also
wanted that feature, and we don't want to add a flag for it, so document
it.

While at it, also document the ways to turn off color entirely. This
used to be just TERM=dumb; we also add NO_COLOR as the opposite to
FORCE_COLOR, which seems to be a standard defined in https://no-color.org/.

We're not updating the tests, because without a real terminal there's no
way to properly test this code, and that simply does not seem worth the
effort for two lines of simple logic. Extending the code with mocks so
we can test without a real terminal also does not seem worthwhile.

Fixes #611.