interp: generously increase the Runner.Run test timeout
I was forgetting that we run 'go test -race' on CI on Mac and Windows,
which are somewhat slow and busy machines to begin with. We were already
getting test flakes on master.

Five seconds is generous enough to hopefully not cause more issues, but
it's still short enough to not hang the developer's machine for too
long.