expand: avoid panic when backslash is followed by EOF
It's a rare edge case, but the parser allows it as per the POSIX Shell
spec, so we need to handle it.

Since the behavior of backslash is to escape the following character,
and there is no such following character, we essentially ignore the
backslash entirely, just like Bash.

It's worth noting that POSIX shells like dash and busybox ash expand the
word to "\\" rather than "", though the interpreter tends to behave like
Bash in general:

	$ echo -n 'echo \' | bash

	$ echo -n 'echo \' | dash
	\
	$ echo -n 'echo \' | busybox ash
	\

It doesn't seem like this difference is hugely important, given that all
sane shell scripts should end with a newline.

Fixes #669.