interp: rewrite the scoping of variables
Before, we used the combination of Env with two maps, Vars and cmdVars,
to implement variable scopes. Vars was for regular variables like
"foo=bar", and cmdVars was for inline variables like "foo=bar baz".

This had a significant bug; inline variables can also be scoped, and
since we used a single map for them, an inner scope could modify an
outer scope:

	$ gosh -c 'f() { echo $inline; inline=bar true; echo $inline; }; inline=foo f'
	foo

	$ bash -c 'f() { echo $inline; inline=bar true; echo $inline; }; inline=foo f'
	foo
	foo

We could have fixed this by merging cmdVars into Vars, and fixing up the
restoration of previous values for inline variable assignments.

However, we also noticed that we implemented scoping in two different
ways: one via overlayEnviron, and another via making entire copies of
Vars and cmdVars.

Instead, internally track variables only via the Env field. Since it
must remain an expand.Environ for backwards compatibility, and an
interpreter must also modify variables, we add an unexported writeEnv
field too.

Now, all nested scopes are implemented via overlayEnviron, be it for
function calls, program executions, or any other reason.

Since Vars is an exported field, we can't get rid of it. It could have
been used to query what variables were set after running a program, so
to keep that working, we add a bit of code at the end of Run to fill the
map. Reset always emptied Vars, so it could have never be used to
pre-fill variables; that's what Env is for.

I've also left some v4 thoughts as TODOs. We might want to rethink the
WriteEnviron semantics, or even a larger interface supporting scopes, to
allow the user to control the entire environment behavior.

Fixes #668.