interp: expand array arguments
Bash array values are generally expanded. For example:

```
$ x=($(echo a b)); echo ${x[0]}
a
```

The exception is when an index is provided. Those values are taken
literally.

```
$ x=([0]=$(echo a b)); echo ${x[0]}
a b
```

The runner expands values in `assignVal` and did so by mapping the
expression index to the array index. A fixed mapping fails when a value
generates more than one field. Hence the code interpreted all array
values as literal.

This change iterates over the elements, expanding literally or with
fields, while keeping track of the implicit index. The result is then
flattened down at the computed indexes.

Fixes #671