interp: implicit casting of arrays
`declare a` would create an unset variable of an unknown type. A
variable should be type cast to an array when an array typed assignment
occurs. String typed variables being cast to arrays was covered but
missed the unset case.

A panic was inserted for missing type conversions. All conversion should
be identified, even if not implemented.

Fixes #675