interp: local vars are mutable
Local variables should be mutable but f4c774a accidentally made them
immutable. A new assignment is made with a new `expand.Variable` struct,
which defaults the Local flag to false. Failing to preserve the
localness of an overlayed variable means future sets are incorrectly
forwarded to the global scope.

Fixes #674