cmd/shfmt: continue when a path results in an error
We used to immediately return and stop if any path resulted in an error.
This is not correct behavior; if the first path does not exist, we
should still format a second path.

This was made worse by how non-empty diffs are internally treated as
errors, so we'd always stop the entire program when a file resulted in a
diff.

This is a regression introduced in v3.2.0 when we overhauled the walking
code. The cause was that we never had tests with multiple files. We do
now.

Fixes #682.