add Bash 5.1 parameter transformations, tweak test cases
Per the NEWS:

dd. New `U', `u', and `L' parameter transformations to convert to uppercase,
    convert first character to uppercase, and convert to lowercase,
    respectively.

hh. New `K' parameter transformation to display associative arrays as key-
    value pairs.

In my testing, mksh only supports ${a@Q} and ${a@#}, and does not support
references or arrays as the parameter. So I have added tests to ensure:

  1. Q works in mksh and bash
  2. # only works in mksh
  3. a, u, A, E, K, L, P, and U only work in bash.

I also modified the file tests to reflect the above support.

Note that bash does not report an error parsing expansions it does not support.