move fuzz.go out of the root directory
Go 1.16.x and older versions complain when using "go get" to update the
module:

	$ go get -d mvdan.cc/sh/v3@latest
	mvdan.cc/sh/v3: no Go source files

This is because the root directory looks like a package, having one Go
file, but it's actually ignored via the gofuzz build tag.

The upstream bug is fixed in the upcoming Go 1.17, but users will
continue relying on 1.16.x for at least half a year.

There's an easy solution: since we're only fuzzing the syntax package,
move the go-fuzz files there.

Fixes #703.