syntax: add a Quote func
Similar to strconv.Quote, but with shell syntax.
See the added godoc for more details.

Unlike strconv.Quote, it needs to return a boolean,
as not all strings can be represented in shell.

With an example to show and test the basics,
as well as a fuzz function to verify edge cases.

Fixes #328.