syntax,expand: handle null bytes like bash
When lexing, skip them.

When expanding, either skip them or treat them as the end of the string.
This, again, is matching bash's behavior.

This seems to stem from the fact that strings in shell are C strings,
which are null-terminated.
Go strings aren't null-terminated, and can contain null bytes.
We need to optimize for compatibility over performance.