syntax: convert the parser+printer fuzzer to testing.F
We already use Go 1.18's fuzzing support for FuzzQuote,
and since it's working well, do it here too.

It supports structured input parameters,
which will mean being able to get rid of the bit math.
For now, just split the []byte into three components,
and add a TODO for splitting those up further.

Finally, we add a TODO about the initial corpus.
With go-fuzz, we relied on a stored corpus in a git branch.
With testing.F, we'll need something slightly different.