interp: make creation of fifos much more robust
CI ran into a flake again on Linux,
and while investigating that locally,
it dawned on me why process substitutions are flaky.
The added comment explains why.

With the old code, the command below ran into an error within 20s:

    go test -short -count=1000 -failfast

With the new code, I ran it for multiple minutes with no errors.
I also verified that the retry mechanism works,
by dropping the random part back to a uint32 to trigger more retries.

Updates #513.